refactor: move higher redisconnector logic to hli
