+++ bec_lib/tests/test_redis_connector_fakeredis.py
@@ -21,7 +21,7 @@ from bec_lib.redis_connector import (
 from bec_lib.redis_connector.buffered_redis_connector import BufferedRedisConnector
 from bec_lib.serialization import MsgpackSerialization
 
-from .test_redis_connector import TestMessage
+from .test_buffered_redis_connector import TestMessage
 
 # pylint: disable=protected-access
 # pylint: disable=missing-function-docstring
@@ -627,39 +627,6 @@ def test_lrem(connected_connector: BufferedRedisConnector):
     conn.lrem(ep(msgs[1].queue), 0, msgs[1])
     list_contents = conn.lrange(ep("primary"), 0, -1)
     assert list_contents == [msgs[0], msgs[2]]
-
-
-def test_connector_publish_metrics(connected_connector):
-
-    start = time.time()
-    data = []
-    ep = MessageEndpoints.dynamic_metric("test")
-
-    def cb(msg):
-        nonlocal data
-        data.append(msg.value)
-
-    connected_connector.register(ep, cb=cb, start_thread=False)
-    connected_connector.publish_metrics(
-        "test",
-        {
-            "m1": 5,
-            "m2": 5.5,
-            "m3": {"value": "test", "possible_values": ["test", "prod"]},
-            "m4": True,
-        },
-    )
-    connected_connector.poll_messages(timeout=1)
-
-    stop = time.time()
-    res = data[0]
-    assert isinstance(res, messages.DynamicMetricMessage)
-    assert start <= res.timestamp <= stop
-    assert res.metrics["_m1"].value == 5
-    assert res.metrics["_m2"].value == 5.5
-    assert res.metrics["_m3"].value == "test"
-    assert set(res.metrics["_m3"].possible_values) == set(["prod", "test"])
-    assert res.metrics["_m4"].value is True
 
 
 def test_merging_streams_does_not_skip_messages(connected_connector: BufferedRedisConnector):
